refactor partials to load individual stylesheet - main layout now adds dynamic link to partial-specific stylesheet - refactors fluent.js to expose `getLocale()` using AsyncLocalStorage
