+++ src/client/css/index.css
@@ -2,4 +2,3 @@
 @import './variables.css';
 @import './global.css';
 @import './nav.css';
-@import './landing.css';
