+++ src/controllers/landing.js
@@ -1,9 +1,8 @@
-import { mainLayout } from '../views/layouts/main.js'
+import { mainLayout } from '../views/main.js'
 import { landing } from '../views/partials/landing.js'
 
 function landingPage (req, res) {
   const data = {
-    locale: req.locale,
     partial: landing
   }
